requirement.txt
streamlit
mediapipe==0.10.11
opencv-python-headless
numpy
Pillow
requests